Removed code associated with the transaction_list module and its endpoint
