+++ src/services/mod.rs
@@ -7,7 +7,6 @@ pub mod hooks;
 pub mod safes;
 pub mod transactions_details;
 pub mod transactions_history;
-pub mod transactions_list;
 pub mod transactions_proposal;
 pub mod transactions_queued;
